Building JSON works
JSON seems correct (has been tested) , there is a file if u wanna check.
Reworked the launcher a little bit (I do not know what  Arturo did in there). We should divide the Controller of GUI and the controller of Console so it is more fluent and less messy, will say it on discord.
Solved a bug of abilityCommand that made the problem explode.