Various changes in game
Ordered class game and discovered several methods that were either uncalled, not useful, a repetition of other methods or that assumed responsabilities that were not theirs, the rest of the changes in other classes are the consequence of renaming and refactoring this methods.

It has been tested with a normal match and worked fine, hope I did not miss anything.