God, in his infinite wisdom, has a plan for us all
And so I say: Live your life, love the people in it, marvel at the beauty of his creation. Do that which truly makes you happy. For as long as he is in your heart, you are always following his plan.

Which is why I proudly present a new and painful refactor. Because nothing makes me happier than writing bad code